Finished up to and including section 8.
More support for including extra files. This was prompted by the endless errors
from err_sys not being present when altering some of the given code snippets
and not compiling them using the included makefiles. Definitely a hack, might be
better to start modifying and using the makefiles for these kinds of exercises.